pyproject.toml
[project]
name = "deep-knowledge-agent"
version = "0.1.0"
description = "Deep Knowledge Agent - AI-powered research assistant for in-depth analysis, knowledge synthesis, and comprehensive research on complex topics"
authors = [{ name = "Paras Chamoli", email = "paras25619@gmail.com" }]
readme = "README.md"
keywords = ['python', 'ai', 'agent', 'research', 'knowledge', 'agno', 'bindu']
requires-python = ">=3.12,<3.13"
dependencies = [
    "bindu==2026.1.12",
    "agno>=2.2.0",
    "fastmcp>=2.11.0",
    "requests>=2.31.0",
    "rich>=13.0.0",
    "openai>=2.11.0",
    "sqlalchemy>=2.0.45",
    "mem0ai>=1.0.1",
    "lancedb>=0.8.0",
    "tantivy>=0.22.0",
    "inquirer>=3.2.0",
    "typer>=0.12.0",
    "python-dotenv>=1.0.0",
]

[project.urls]
Homepage = "https://github.com/Paraschamoli/deep-knowledge-agent"
Repository = "https://github.com/Paraschamoli/deep-knowledge-agent"
Documentation = "https://github.com/Paraschamoli/deep-knowledge-agent#readme"

[dependency-groups]
dev = [
    "pytest>=7.2.0",
    "pytest-asyncio>=0.23.0",
    "pre-commit>=2.20.0",
    "tox-uv>=1.11.3",
    "ty>=0.0.1a16",
    "pytest-cov>=4.0.0",
    "ruff>=0.11.5",
    "mkdocs>=1.4.2",
    "mkdocs-material>=8.5.10",
    "mkdocstrings[python]>=0.26.1",
    "types-requests>=2.31.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["deep_knowledge_agent"]

[tool.ty.environment]
python = "./.venv"
python-version = "3.12"

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[tool.ruff]
target-version = "py312"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    "YTT",
    "S",
    "B",
    "A",
    "C4",
    "T10",
    "SIM",
    "I",
    "C90",
    "E", "W",
    "F",
    "PGH",
    "UP",
    "RUF",
    "TRY",
]
ignore = [
    "E501",
    "E731",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["S101"]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["deep_knowledge_agent"]
